% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_smos.R
\name{missing_smos}
\alias{missing_smos}
\title{Print the dates for which BEC-SMOS soil moisture data have not been found}
\usage{
missing_smos()
}
\value{
a character string containing dates for which the data files were not
found on the server.
}
\description{
This function prints out the dates for which BEC-SMOS soil moisture data with
specified \code{frequency} and \code{orbit} arguments have not been found on
the BEC server. This information is automatically generated while running
\code{\link[=find_smos]{find_smos()}}, but displayed only if requested by the
user.
}
\examples{
\dontrun{
missing_smos()
}

}
\references{
Pablos M, Gonzalez-Haro C, Portal G, Piles M, Vall-llossera M,
Portabella M (2022). SMOS L4 Surface Soil Moisture downscaled maps at 1 km
EASE-2 (reprocessed mode) (V.6.0) [Dataset].
}
