% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raster_smos.R
\name{plot_raster_smos}
\alias{plot_raster_smos}
\title{Draw a raster image of BEC-SMOS soil moisture data}
\usage{
plot_raster_smos(data, lat = NULL, lon = NULL, QA = NULL)
}
\arguments{
\item{data}{a character string containing a link to a single BEC-SMOS data
file stored on the local computer.}

\item{lat}{a numeric vector of length 2 containing latitudinal bounds of
the plotting region (in 'latlon' projection). Default value is \code{NULL}
which means all data between min and max latitudes are drawn.}

\item{lon}{a numeric vector of length 2 containing longitudinal bounds of
the plotting region (in 'latlon' projection). Default value is \code{NULL}
which means all data between min and max longitudes are drawn.}

\item{QA}{a numeric vector specifying the desired quality assurance of the
data to be drawn. Possible values range from 0 (good quality data) to 15.
To know the meanings of QA > 0, please refer to the technical note on the
BEC-SMOS soil moisture products available at
\url{https://digital.csic.es/handle/10261/303808}.}
}
\description{
This function draws a raster image of BEC-SMOS soil moisture data
corresponding to a single data file and specific geographical extent.
}
\details{
This function reads an original BEC-SMOS soil moisture data file in NetCDF
format, converts data from EASE-2 grid cells to geographic coordinates, and
draws a raster image of soil moisture estimates in 'latlon' projection. The
image can be drawn for a specific geographical extent if requested by the
user. Otherwise, the entire dataset across Europe (between 28 and 72 degrees
 north and -11 and 40 degrees east) will be plotted.

Note that due to high resolution of the data (~1 km), the execution of this
function may take a long time to be completed depending on the amount of
data to be drawn.
}
\examples{
\dontrun{
# to draw a raster image of soil moisture data within specified bounds
smos_data <- list_smos()
lat <- c(35.00, 45.00)
lon <- c(-10.50, 4.50)
plot_raster_smos(data = smos_data[1], lat = lat, lon = lon)
}

}
