% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_smos.R
\name{extract_smos}
\alias{extract_smos}
\title{Extract BEC-SMOS soil moisture estimates for specific geographical locations}
\usage{
extract_smos(data, lat, lon, save = FALSE, filename = "SM_output.csv")
}
\arguments{
\item{data}{a character vector as produced by \code{list_smos()} containing
links to the data files on the local computer.}

\item{lat}{a numeric vector containing latitudes of points to extract the
data from (in 'latlon' projection).}

\item{lon}{a numeric vector containing longitudes of points to extract the
data from (in 'latlon' projection).}

\item{save}{a logical vector indicating whether the output should  be saved
as a CSV file. Default is \code{FALSE}.}

\item{filename}{a character string naming a file for saving the output. If
\code{save} = \code{TRUE} and no \code{filename} is specified by the user,
the data is saved in a file with a generic name 'SM_output.csv'.}
}
\value{
a data.matrix with the relevant information as described in Details.
}
\description{
This function facilitates reading the original BEC-SMOS soil moisture data
files and extracting relevant information for specific geographical locations
by using Lat/Lon coordinates in decimal degrees.
}
\details{
This function reads the original BEC-SMOS soil moisture data files in NetCDF
format, converts data from EASE-2 grid cells to geographic coordinates, and
extracts relevant information for Lat/Lon locations specified by the user.

The data retrieved from each data file includes:

- the coordinates of spatial points (Lon and Lat) from which the data were
extracted;

- frequency and SMOS orbit of each file over which the function iterated;

- date and time when the data was obtained;

- soil moisture estimate (SM);

- quality assurance (QA) corresponding to each SM estimate. Good quality data
is marked with QA = 0. To know the meanings of QA > 0, please refer to the
technical note on the BEC-SMOS soil moisture products available at
\url{https://digital.csic.es/handle/10261/303808}.

The output of this function could be saved as a CSV file.
}
\examples{
\dontrun{
# to iterate over BEC-SMOS data files stored in the current working directory
# and extract soil moisture estimates for the specified geographical locations
smos_data <- list_smos()
lat <- c(40.42, 41.90, 48.86, 52.50, 59.91)
lon <- c(-3.70, 12.50, 2.35, 13.40, 10.75)
sm_estimates <- extract_smos(data = smos_data, lat = lat, lon = lon)
}

}
